Get kind mgmt cluster logs in clusterctl upgrade test
